pyproject.toml
[tool.poetry]
name = "todo-project"
version = "0.1.0"
description = ""
authors = ["PedroSmaxY <pedrohnovais@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
drf-nested-routers = "^0.94.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
